move to UA handling to device file
Move the UA code so it can be called from other files.
