+++ libtcmu_common.h
@@ -33,6 +33,8 @@ enum {
 	TCMU_STS_NO_RESOURCE,
 	/* handler has setup sense. */
 	TCMU_STS_PASSTHROUGH_ERR,
+	/* UA code has setup the sense */
+	TCMU_STS_UNIT_ATTENTION,
 	TCMU_STS_BUSY,
 	TCMU_STS_WR_ERR,
 	TCMU_STS_RD_ERR,
@@ -46,7 +48,6 @@ enum {
 	TCMU_STS_HW_ERR,
 	TCMU_STS_RANGE,
 	TCMU_STS_FRMT_IN_PROGRESS,
-	TCMU_STS_CAPACITY_CHANGED,
 	TCMU_STS_NOTSUPP_SAVE_PARAMS,
 	TCMU_STS_WR_ERR_INCOMPAT_FRMT,
 	TCMU_STS_TRANSITION,
